Support JSON and Markdown reports in addition to HTML
This change adds two extra report formats, JSON and Markdown, in
addition to the HTML reports already available. The Markdown
report is based on support for Markdown reports already present in
Xcov.

Three new fastlane / command line options allow the user to choose
which format or formats they want their reports in. To retain
backwards compatibility and keep in the spirit of Xcov's no-hassle
reports, the HTML report is on by default and the JSON and Markdown
reports are off.