Make JSON report honour ignored files in coverage
The current JSON report just spits out the raw report received
from xcov-core. As such, the coverage total, and the coverage
per target values in the JSON report are different from those in
the HTML or Markdown reports.

This patch creates a new JSON report using the same data that is
used for the HTML and Markdown reports. Targets which are skipped
(such as test targets) are not included in the output, and files
that have been ignored are marked with "ignored": true in the file
object and are not considered when calculating per target or per
report totals.