Add include targets option
With Xcode 8 it seems that coverage is generated for all targets in an app even it the schem for that target has coverage turned off.
When there are a lot of targets, for exampl projects that use cocoapods, an inclusion list is much easier to maintain than an exclusion list.

With this change a list of targets to include in the report can be specified as well as the exclusion list that existed previously.