Update gemspec to support fastlane 2.0
With the recent [mono gem release](https://github.com/fastlane/fastlane/releases/tag/2.0.0) you'll need to remove _fastlane_core_ as dependency. 
Sorry I submitted this PR on such a short notice, I submitted PRs on various projects, but for some reasons I forgot to do it for this widely used one. 
Please give this a try with _fastlane_ 2.0, make sure everything works as usual and push a new release  👍
Thanks for building this!