Fix xccov merge error if processing multiple xcresult files
Apple's xccov tool errors out if a file with the same name exists, but
handling that for our file `merged.xccovreport` isn't enough as that
tool also creates a directory called `merged.xccovarchive` - this one
with a name chosen by the tool, not by us - so regardless of the name we
choose for the xccov report, this archive would always be the same,
leading to failures on subsequent runs.
This commit avoids any clashes by removing existing files and renaming
Apple-created directory to a more unique name before the next iteration.